+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Support disabling automatic viewport interaction with `neoscore.show(auto_viewport_interaction_enabled=False)`. This disables scroll-zooming, drag-moving, and the appearance of window scrollbars.
 - Support setting the preview window size with new `neoscore.show()` options `min_window_size` and `max_window_size`.
 - Support launching the preview window in fullscreen mode with `neoscore.show(fullscreen=True)`.
+- Support programmatically controlling the viewport's center position, scale (zoom), and rotation with `neoscore.set_viewport_center_pos()`, `neoscore.set_viewport_scale()`, and `neoscore.set_viewport_rotation()`. Corresponding getters are also provided.
 
 # 0.1.1 (2022-05-20)